// src-ui/core/ipc/generated/BotBasicConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutoRestartSchedule } from "./AutoRestartSchedule";
import type { BackendType } from "./BackendType";

export type BotBasicConfig = { name: string, QQID: number, musicSignUrl: string, autoRestartSchedule: AutoRestartSchedule, offlineAutoRestart: boolean, runtime_target: string, backend_type: BackendType, };